Migrate away from clap macro and use builder pattern
Using the clap macro was a mistake. It's a whole other macro to use with
pretty poor documentation. This is now much easier to use and
understand. If you need a new feature, you can just call the right
method and not try to guess how to use that with the macro.

It also allows us to use long flags with `-` in the middle of them